Gemma4 阶段F：分层组 Paged KV Cache + parity 通过
- 新增 Gemma4PagedKVCacheManager：按 (kv_heads, head_dim) 分 cache group
  （D=256 滑动组 12 层 / D=512 full 组 3 层），组内物理张量含 layer 维
  [n_group_layers, n_blocks, block_size, kv, dim]；共享层 15-34 不占存储，
  读源层 block；block_table 跨组共享；gather 用 advanced indexing +
  单条目 index memo（每步 15 层复用一次索引计算）
- 引擎 runner 统一 write_kv_for_tokens/gather 接口，密集版管理器退役
- 修复 RoPE dtype：cos/sin 按官方语义转 x.dtype 再乘，避免 fp32 混入 bf16 cache
- 修复 prefill 误用窗口切片导致 K/V 与 mask 错位（切片仅用于 decode 单 query）
- 真模型 MPS bf16：paged 引擎与直连参考 24 token 逐 token parity OK

+++ miniservellm/runtime/gemma4_runner.py
@@ -129,7 +129,11 @@ class Gemma4EagerTextRunner:
         return emb.cos().to(torch.float32), emb.sin().to(torch.float32)
 
     def _apply_rope(self, x: torch.Tensor, cos: torch.Tensor, sin: torch.Tensor) -> torch.Tensor:
-        # x: [T, H, D]; cos/sin: [T, D]
+        # x: [T, H, D]; cos/sin: [T, D]. Official implementation casts
+        # cos/sin to x.dtype before applying (modeling_gemma4.py rotary
+        # forward), keeping q/k in the model dtype.
+        cos = cos.to(x.dtype)
+        sin = sin.to(x.dtype)
         return x * cos[:, None, :] + _rotate_half(x) * sin[:, None, :]
 
     def _mask(self, attention_type: str, seq_len: int, dtype: torch.dtype) -> torch.Tensor:
